Added Gender, Birthday, Death day, OfficialSite attributes

+++ TvMaze/Domain/Rating.cs
@@ -1,5 +1,5 @@
 ﻿namespace TvMaze.Domain {
     public class Rating {
-        public float? Average { get; set; }
+        public decimal? Average { get; set; }
     }
 }
