+++ TvMaze/Domain/SearchResult.cs
@@ -1,6 +1,6 @@
 ﻿namespace TvMaze.Domain {
     public class SearchResult<T> where T:class {
-        public float Score { get; set; }
+        public decimal Score { get; set; }
         public T Element { get; set; }
     }
 }
